requirements.txt
fastapi
uvicorn[standard]
httpx
pytz
supabase
python-dotenv
openai
